requirements-dev.txt
aioredis==1.0.0
asgiref==2.3.2
async-generator==1.9
async-timeout==2.0.1
attrs==17.4.0
autobahn==18.4.1
Automat==0.6.0
channels==2.1.5
channels-redis==2.1.1
constantly==15.1.0
daphne==2.2.3
Django==2.1.7
django-cors-headers==2.2.0
djangorestframework==3.8.2
dry-rest-permissions==0.1.10
Faker==0.8.12
freezegun==0.3.10
gunicorn==19.9.0
hiredis==0.2.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
mock==2.0.0
more-itertools==4.1.0
msgpack==0.5.6
pluggy==0.6.0
py==1.5.3
pytest==3.5.0
pytest-asyncio==0.8.0
pytest-django==3.1.2
pytest-env==0.6.2
pytest-mock==1.10.0
python-dateutil==2.7.2
pytz==2018.4
six==1.11.0
Sphinx==1.7.5
text-unidecode==1.2
tornado==5.0.2
Twisted==18.7.0
txaio==2.10.0
uvicorn==0.6.1
zope.interface==4.4.3
